Render ajax modal and refactor
- Render ajax modal to use for paginated faceted multiselect
- Refactored the request builder for index/count
