+++ app/assets/javascripts/multifacets.js
@@ -5,3 +5,5 @@ $(document).ready(function(){
 function resetFacets(){
   $("input[type='hidden'][name='f_inclusive[digital_collection_sim][]']").remove();
 }
+
+Blacklight.ajaxModal.triggerLinkSelector  += ",a.more_multiselect_facets_link";
