+++ app/controllers/catalog_controller.rb
@@ -280,4 +280,22 @@ class CatalogController < ApplicationController
     solr_parameters[:fq]
   end
 
+  # displays values and pagination links for a single facet field
+  def multiselect_facet
+    @facet = blacklight_config.facet_fields[params[:id]]
+    @response = get_facet_field_response(@facet.field, params)
+    @display_facet = @response.facets.first
+
+    @pagination = facet_paginator(@facet, @display_facet)
+
+
+    respond_to do |format|
+      # Draw the facet selector for users who have javascript disabled:
+      format.html
+      format.json { render json: render_facet_list_as_json }
+
+      # Draw the partial for the "more" facet modal window:
+      format.js { render :layout => false }
+    end
+  end
 end
