+++ app/views/catalog/_multiselect_facet_pagination.html.erb
@@ -1,0 +1,22 @@
+<span>_multiselect_facet_pagination</span>
+    <div class="prev_next_links btn-group pull-left">
+      <%= link_to_previous_page @pagination, raw(t('views.pagination.previous')), :params => params, :param_name => Blacklight::Solr::FacetPaginator.request_keys[:page], :class => 'btn btn-link', :data => {:ajax_modal => "preserve"} do %>
+       <span class="disabled btn btn-disabled"><%= raw(t('views.pagination.previous')) %></span>
+     <% end %>
+      <%= link_to_next_page @pagination, raw(t('views.pagination.next')), :params => params, :param_name => Blacklight::Solr::FacetPaginator.request_keys[:page], :class => 'btn btn-link',  :data => {:ajax_modal => "preserve"} do %>
+
+       <span class="disabled btn  btn-disabled"><%= raw(t('views.pagination.next')) %></span>
+      <% end %>
+
+    </div>
+
+  <div class="sort_options btn-group pull-right">
+    <% if @pagination.sort == 'index' -%>
+      <span class="active az btn btn-default"><%= t('blacklight.search.facets.sort.index') %></span><%= link_to_unless(@pagination.sort == 'count', t('blacklight.search.facets.sort.count'),
+        @pagination.params_for_resort_url('count', params), :class => "sort_change numeric btn btn-default",
+        :data => {:ajax_modal => "preserve"}) %>
+    <% elsif @pagination.sort == 'count' -%>
+      <%=  link_to(t('blacklight.search.facets.sort.index'), @pagination.params_for_resort_url('index', params),
+        :class => "sort_change az btn btn-default",  :data => {:ajax_modal => "preserve"}) %><span class="active numeric btn btn-default"><%= t('blacklight.search.facets.sort.count') %></span>
+    <% end -%>
+  </div>
