+++ app/views/catalog/multiselect_facet.html.erb
@@ -1,10 +1,10 @@
 <div class="facet_pagination top">
-  <%= render :partial=>'facet_pagination' %>
+  <%= render :partial=>'multiselect_facet_pagination' %>
 </div>
 
 <div class="modal-header">
   <button type="button" class="ajax-modal-close close" data-dismiss="modal" aria-hidden="true">×</button>
-  <h3 class="modal-title"><%= facet_field_label(@facet.key) %></h3>
+  <h3 class="modal-title"><%= facet_field_label(@facet.field) %></h3>
 </div>
 <div class="modal-body">
   <div class="facet_extended_list">
